+++ packages/blade/src/components/ActionList/ActionListItem.tsx
@@ -17,11 +17,11 @@ import { useDropdown } from '~components/Dropdown/useDropdown';
 import type { Feedback } from '~tokens/theme/theme';
 import { Text } from '~components/Typography';
 import { isReactNative, makeAccessible, makeSize, metaAttribute, MetaConstants } from '~utils';
-import type { WithComponentId } from '~utils';
 import { Checkbox } from '~components/Checkbox';
 import size from '~tokens/global/size';
 import type { DropdownProps } from '~components/Dropdown';
 import type { StringChildrenType, TestID } from '~src/_helpers/types';
+import { assignWithoutSideEffects } from '~src/utils/assignWithoutSideEffects';
 
 type ActionListItemProps = {
   title: string;
@@ -99,12 +99,12 @@ type ActionListSectionProps = {
    */
   _hideDivider?: boolean;
 } & TestID;
-const ActionListSection: WithComponentId<ActionListSectionProps> = ({
+const _ActionListSection = ({
   title,
   children,
   testID,
   _hideDivider,
-}): JSX.Element => {
+}: ActionListSectionProps): JSX.Element => {
   return (
     <BaseBox
       {...makeAccessible({
@@ -133,9 +133,11 @@ const ActionListSection: WithComponentId<ActionListSectionProps> = ({
   );
 };
 
-ActionListSection.componentId = componentIds.ActionListSection;
+const ActionListSection = assignWithoutSideEffects(_ActionListSection, {
+  componentId: componentIds.ActionListSection,
+});
 
-const ActionListItemIcon: WithComponentId<{ icon: IconComponent }> = ({ icon }): JSX.Element => {
+const _ActionListItemIcon = ({ icon }: { icon: IconComponent }): JSX.Element => {
   const Icon = icon;
   const { intent, isDisabled } = React.useContext(ActionListItemContext);
   return (
@@ -150,9 +152,15 @@ const ActionListItemIcon: WithComponentId<{ icon: IconComponent }> = ({ icon }):
   );
 };
 
-ActionListItemIcon.componentId = componentIds.ActionListItemIcon;
+const ActionListItemIcon = assignWithoutSideEffects(_ActionListItemIcon, {
+  componentId: componentIds.ActionListItemIcon,
+});
 
-const ActionListItemText: WithComponentId<{ children: StringChildrenType }> = ({ children }) => {
+const _ActionListItemText = ({
+  children,
+}: {
+  children: StringChildrenType;
+}): React.ReactElement => {
   const { isDisabled } = React.useContext(ActionListItemContext);
 
   return (
@@ -162,7 +170,9 @@ const ActionListItemText: WithComponentId<{ children: StringChildrenType }> = ({
   );
 };
 
-ActionListItemText.componentId = componentIds.ActionListItemText;
+const ActionListItemText = assignWithoutSideEffects(_ActionListItemText, {
+  componentId: componentIds.ActionListItemText,
+});
 
 const ActionListCheckboxWrapper = styled(BaseBox)<{ hasDescription: boolean }>((_props) => ({
   pointerEvents: 'none',
@@ -367,9 +377,9 @@ const _ActionListItem = (props: ActionListItemProps): JSX.Element => {
   );
 };
 
-const ActionListItem = React.memo(_ActionListItem);
-// @ts-expect-error
-ActionListItem.componentId = componentIds.ActionListItem;
+const ActionListItem = assignWithoutSideEffects(React.memo(_ActionListItem), {
+  componentId: componentIds.ActionListItem,
+});
 
 export {
   ActionListItem,