+++ packages/blade/src/components/Button/BaseButton/BaseButton.tsx
@@ -46,6 +46,7 @@ import type { BaseTextProps } from '~components/Typography/BaseText/types';
 import type { BladeElementRef } from '~src/hooks/useBladeInnerRef';
 import { useBladeInnerRef } from '~src/hooks/useBladeInnerRef';
 import { getStringFromReactText } from '~src/utils/getStringChildren';
+import { assignWithoutSideEffects } from '~src/utils/assignWithoutSideEffects';
 
 type BaseButtonCommonProps = {
   size?: 'xsmall' | 'small' | 'medium' | 'large';
@@ -453,7 +454,8 @@ const _BaseButton: React.ForwardRefRenderFunction<BladeElementRef, BaseButtonPro
   );
 };
 
-const BaseButton = React.forwardRef(_BaseButton);
-BaseButton.displayName = 'BaseButton';
+const BaseButton = assignWithoutSideEffects(React.forwardRef(_BaseButton), {
+  displayName: 'BaseButton',
+});
 
 export default BaseButton;