+++ packages/blade/src/components/Link/BaseLink/__tests__/BaseLink.native.test.tsx
@@ -75,6 +75,14 @@ describe('<BaseLink />', () => {
     expect(toJSON()).toMatchSnapshot();
   });
 
+  it('should add hitSlop to link', () => {
+    const linkText = 'Learn More';
+    const hitSlop = { top: 4, right: 2, bottom: 4, left: 2 };
+    const { getByRole } = renderWithTheme(<BaseLink hitSlop={hitSlop}>{linkText}</BaseLink>);
+    const link = getByRole('link');
+    expect(link.findByProps({ hitSlop })).toBeTruthy();
+  });
+
   it('should render button variant of link', () => {
     const linkText = 'Learn More';
     const { toJSON, getByRole } = renderWithTheme(<BaseLink variant="button">{linkText}</BaseLink>);